Add GameActivity/Gradle project support (#98)
* Change project structure

* Add platform android

* Add gradlew and App

* Add availability to compile rust library without build apk

* Add app gradle project

* Working NA with Macroquad

* Fix clippy and fmt

* Add permission result

* Move permissions to crossbow-android

* Alter permissions

* Add generation of gradle project with flag

* Refactor target dir for AAB

* Fix tests

* Run cargo clippy

* Fix fmt

* Fix gradle generation

* Add integration test for gradle project

* Alter to use Kotlin and fix bug with Activity not found

* Add android metadata to cargo manifest

* Fix cargo issue

* Add run gradle command

* Alter generation of gradle to use :app

* Add export to directory flag option

* Alter CrossbowApp AndroidManifest.xml, fix issues and bump versions

+++ crossbundle/tools/src/commands/android/helper_functions.rs
@@ -3,31 +3,12 @@ use crate::error::*;
 /// Helper function to delete files
 pub fn remove(target: Vec<std::path::PathBuf>) -> Result<()> {
     target.iter().for_each(|content| {
-        if content.is_file() {
+        if content.exists() && content.is_file() {
             std::fs::remove_file(&content).unwrap();
         }
-        if content.is_dir() {
+        if content.exists() && content.is_dir() {
             std::fs::remove_dir_all(&content).unwrap();
         }
     });
     Ok(())
-}
-
-/// Helper function to parse the file to slice u8 from assets folder.
-pub fn parse_the_file_to_slice_u8(
-    file_name: &str,
-    app_name: &str,
-    additional_dir: Option<String>,
-) -> Result<Vec<u8>> {
-    let current_dir = std::env::current_dir()?;
-    let mut buf = current_dir;
-    buf.push("examples");
-    buf.push(app_name);
-    buf.push("assets");
-    if let Some(add) = additional_dir {
-        buf.push(add);
-    }
-    buf.push(file_name);
-    let bytes = std::fs::read(buf)?;
-    Ok(bytes)
 }
